+++ CHANGELOG.md
@@ -26,6 +26,7 @@ For more information on SemVer, please visit http://semver.org/.
 * Added collapsable filedset in Controls Layout
 * Added "fonts" component
 * Added `wpdk_page_with_slug()` function
+* Improved and fixed `WPDKWidget` css inline for "customize.php"
 * Improved compatibility with "costomize.php" page
 * Improved and cosmetic on file `classes/core/wpdk-functions.php`
 * Fixed several Javascript events - add "off" where needed